requirements.txt
SQLAlchemy~=2.0.31
openai~=1.35.14
python-dotenv~=1.0.1
langgraph~=0.1.9
langchain~=0.2.9
streamlit~=1.36.0